style: opt style about testing connection
