+++ app/src/main/java/io/github/trojan_gfw/igniter/MainActivity.java
@@ -7,7 +7,6 @@ import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
 import android.net.VpnService;
-import android.os.AsyncTask;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v4.content.LocalBroadcastManager;
@@ -18,7 +17,6 @@ import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Switch;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import org.json.JSONObject;
 
@@ -26,11 +24,6 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.InputStream;
-import java.lang.ref.WeakReference;
-import java.net.InetSocketAddress;
-import java.net.Proxy;
-import java.net.URL;
-import java.net.URLConnection;
 
 
 public class MainActivity extends AppCompatActivity {
@@ -105,72 +98,6 @@ public class MainActivity extends AppCompatActivity {
         clashLink.setEnabled(inputEnabled);
     }
 
-    private static class TestConnectionResult {
-        final String url;
-        final boolean isConnected;
-        final Exception error;
-        final long time; // In milliseconds
-
-        TestConnectionResult(String url, boolean isConnected, Exception error, long time) {
-            this.url = url;
-            this.isConnected = isConnected;
-            this.error = error;
-            this.time = time;
-        }
-    }
-
-    private static class TestConnection extends AsyncTask<String, Void, TestConnectionResult> {
-        private WeakReference<MainActivity> activityReference;
-        TestConnection(MainActivity context) {
-            activityReference = new WeakReference<>(context);
-        }
-        protected void onPreExecute() {
-            super.onPreExecute();
-            MainActivity activity = activityReference.get();
-            if (activity != null) {
-                activity.testConnectionButton.setText(R.string.testing);
-                activity.testConnectionButton.setEnabled(false);
-            }
-        }
-
-        protected TestConnectionResult doInBackground(String... urls) {
-            String url = urls[0];
-            try {
-                long startTime = System.currentTimeMillis();
-                ProxyService serviceInstance = ProxyService.getInstance();
-                InetSocketAddress proxy_address = new InetSocketAddress("127.0.0.1",
-                        (int)serviceInstance.tun2socksPort);
-                Proxy proxy = new Proxy(Proxy.Type.SOCKS, proxy_address);
-                URLConnection connection = new URL(url).openConnection(proxy);
-                connection.setConnectTimeout(1000 * 10); //10 seconds
-                connection.connect();
-                return new TestConnectionResult(url, true, null,
-                        System.currentTimeMillis() - startTime);
-            } catch (Exception e) {
-                return new TestConnectionResult(url, false, e, 0);
-            }
-        }
-
-        protected void onPostExecute(TestConnectionResult result) {
-            MainActivity activity = activityReference.get();
-            if (activity != null) {
-                activity.testConnectionButton.setText(R.string.test_connection);
-                activity.testConnectionButton.setEnabled(true);
-                if (result.isConnected) {
-                    Toast.makeText(activity,
-                            activity.getString(R.string.connected_to__in__ms,
-                                    result.url, String.valueOf(result.time)),
-                            Toast.LENGTH_LONG).show();
-                } else {
-                    Toast.makeText(activity,
-                            activity.getString(R.string.failed_to_connect_to__,
-                                    result.url, result.error.getMessage()),
-                            Toast.LENGTH_LONG).show();
-                }
-            }
-        }
-    }
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -235,7 +162,6 @@ public class MainActivity extends AppCompatActivity {
         };
 
     }
-
 
     @Override
     protected void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
